\name{CardinalitySort}
\alias{CardinalitySort}
\title{Sort a power set by cardinality}
\usage{
CardinalitySort(powerSet)
}
\arguments{
  \item{powerSet}{A power set given by
  \code{\link{PowerSet}}}
}
\value{
powerSet sorted by cardinality
}
\description{
Sorts a power set by the number of elements that it
contains, i.e., its cardinality.
}
\seealso{
\code{\link{order}} that is wrapped by this function
}

