\name{PowerSet}
\alias{PowerSet}
\title{Produces the powerset of a given set. The powerset contains 2^n items
for a set with n items. It includes the empty set and the original set
itself.}
\usage{
PowerSet(set)
}
\arguments{
  \item{The}{set on which to compute the powerset.</param>}
}
\value{
The powerset of the given set. Note that factors are
returned as character
}
\description{
Produces the powerset of a given set. The powerset contains
2^n items for a set with n items. It includes the empty set
and the original set itself.
}
\seealso{
package sets, 2^s can do it, but subset access seems more
complicated
}

